Fix screenshots are being set twice

+++ ClipboardShared/Program.cs
@@ -4,6 +4,7 @@ using System.Threading;
 using System.Windows.Forms;
 using System.Drawing;
 using System.Runtime.InteropServices;
+using System.Drawing.Text;
 
 namespace ClipboardShared
 {
@@ -28,18 +29,6 @@ namespace ClipboardShared
 
     internal class Program
     {
-        private const string c_TextFile = "text.txt";
-        private const string c_ImageFile = "img.png";
-
-        private static event Action ClipboardUpdate;
-
-        private static string directory;
-        private static string textPath;
-        private static string imgPath;
-
-        private static FileSystemWatcher watcher = new FileSystemWatcher();
-        private static int watcherCounter = 0;
-
         private class NotificationForm : Form
         {
             public NotificationForm()
@@ -57,6 +46,19 @@ namespace ClipboardShared
                 base.WndProc(ref m);
             }
         }
+
+        private const string c_TextFile = "text.txt";
+        private const string c_ImageFile = "img.png";
+
+        private static event Action ClipboardUpdate;
+
+        private static string directory;
+        private static string textPath;
+        private static string imgPath;
+
+        private static FileSystemWatcher watcher = new FileSystemWatcher();
+        private static int watcherCounter = 0;
+        private static int clipboardImgCounter = 0;
 
         private static void OnClipboardUpdate()
         {
@@ -93,27 +95,35 @@ namespace ClipboardShared
         private static void ClipboardChangeHandler()
         {
             watcher.EnableRaisingEvents = false; // Don't invoke file changed events for our own change
+            Console.ForegroundColor = ConsoleColor.Green;
             var data = Clipboard.GetDataObject();
             if (data.GetDataPresent(DataFormats.UnicodeText))
             {
                 var contents = data.GetData(DataFormats.UnicodeText) as string;
                 File.WriteAllText(textPath, contents);
+                Console.WriteLine("[Out] text: " + contents);
             }
             else if (data.GetDataPresent(DataFormats.Bitmap))
             {
-                using (MemoryStream pngStream = data.GetData("PNG") as MemoryStream)
+                clipboardImgCounter++;
+                if (clipboardImgCounter == 2)
                 {
-                    if (pngStream != null)
+                    using (MemoryStream pngStream = data.GetData("PNG") as MemoryStream)
                     {
-                        if (File.Exists(imgPath))
+                        if (pngStream != null)
                         {
-                            File.Delete(imgPath);
-                        }
-                        using (FileStream fileStream = new FileStream(imgPath, FileMode.Create))
-                        {
-                            pngStream.WriteTo(fileStream);
+                            if (File.Exists(imgPath))
+                            {
+                                File.Delete(imgPath);
+                            }
+                            using (FileStream fileStream = new FileStream(imgPath, FileMode.Create))
+                            {
+                                pngStream.WriteTo(fileStream);
+                            }
                         }
                     }
+                    Console.WriteLine("[Out] screenshot: " + imgPath);
+                    clipboardImgCounter = 0;
                 }
             }
             watcher.EnableRaisingEvents = true;
@@ -121,40 +131,48 @@ namespace ClipboardShared
 
         private static void FileChangeHandler(Object src, FileSystemEventArgs e)
         {
-            // This event gets fired twice, on start modifying and on finish modifying
-            watcherCounter++;
-            if (watcherCounter == 2)
+            switch (e.Name.ToLower())
             {
-                switch (e.Name.ToLower())
-                {
-                    case c_TextFile:
+                case c_TextFile:
+                    // This event gets fired twice, on start modifying and on finish modifying
+                    watcherCounter++;
+                    if (watcherCounter == 2)
+                    {
                         string text = File.ReadAllText(e.FullPath);
                         if (!string.IsNullOrEmpty(text))
                         {
-                            CallClipboardFun(() => Clipboard.SetText(text));
+                            CallClipboardFun(
+                                "[In] text: " + text,
+                                () => Clipboard.SetText(text)
+                                );
                         }
-                        break;
+                        watcherCounter = 0;
+                    }
+                    break;
 
-                    case c_ImageFile:
-                        using (Bitmap bm = new Bitmap(e.FullPath))
-                        {
-                            CallClipboardFun(() => Clipboard.SetImage(bm));
-                        }
-                        break;
-                }
-                watcherCounter = 0;
+                case c_ImageFile:
+                    using (Bitmap bm = new Bitmap(e.FullPath))
+                    {
+                        CallClipboardFun(
+                            "[In] screenshot: " + e.FullPath,
+                            () => Clipboard.SetImage(bm));
+                    }
+                    break;
             }
         }
 
-        private static void CallClipboardFun(Action action)
+        private static void CallClipboardFun(string msg, Action action)
         {
             // unlisten first, so we don't get invoked by our own change
-            ClipboardUpdate -= ClipboardChangeHandler;
+            ClipboardUpdate = null;
 
             Thread thread = new Thread(action.Invoke);
             thread.SetApartmentState(ApartmentState.STA);
             thread.Start();
             thread.Join();
+
+            Console.ForegroundColor = ConsoleColor.DarkYellow;
+            Console.WriteLine(msg);
 
             ClipboardUpdate += ClipboardChangeHandler;
         }
